updating source code to 1.19
