+++ workshops/minecraft/src-1.11/main/java/org/devoxx4kids/forge/mods/ChatItems.java
@@ -1,0 +1,16 @@
+package org.devoxx4kids.forge.mods;
+
+import net.minecraft.init.Items;
+import net.minecraft.item.ItemStack;
+import net.minecraftforge.event.ServerChatEvent;
+import net.minecraftforge.fml.common.eventhandler.SubscribeEvent;
+
+public class ChatItems {
+
+    @SubscribeEvent
+    public void giveItems(ServerChatEvent event) {
+        if (event.getMessage().contains("potato")) {
+            event.getPlayer().inventory.addItemStackToInventory(new ItemStack(Items.POTATO, 64));
+        }
+    }
+}
