+++ workshops/minecraft/src-1.11/main/resources/mcmod.info
