+++ workshops/minecraft/src/main/java/org/devoxx4kids/forge/mods/DragonSpawner.java
@@ -1,22 +1,50 @@
 package org.devoxx4kids.forge.mods;
 
-import net.minecraft.entity.boss.EntityDragon;
-import net.minecraft.entity.boss.dragon.phase.PhaseList;
-import net.minecraft.init.Blocks;
-import net.minecraft.util.math.BlockPos;
-import net.minecraftforge.event.world.BlockEvent.PlaceEvent;
-import net.minecraftforge.fml.common.eventhandler.SubscribeEvent;
+import net.minecraft.world.entity.EntityType;
+import net.minecraft.world.entity.boss.enderdragon.EnderDragon;
+import net.minecraft.world.entity.boss.enderdragon.phases.EnderDragonPhase;
+import net.minecraft.world.level.block.Blocks;
+import net.minecraftforge.event.level.BlockEvent;
+import net.minecraftforge.eventbus.api.SubscribeEvent;
+import net.minecraftforge.fml.common.Mod;
 
+@Mod.EventBusSubscriber(modid = MainMod.MODID)
 public class DragonSpawner {
-
     @SubscribeEvent
-    public void spawnDragon(PlaceEvent event) {
-        if (event.getPlacedBlock() == Blocks.SPONGE.getBlockState().getBaseState()) {
-            event.getWorld().setBlockToAir(new BlockPos(event.getPos().getX(), event.getPos().getY(), event.getPos().getZ()));
-            EntityDragon dragon = new EntityDragon(event.getWorld());
-            dragon.setLocationAndAngles(event.getPos().getX(), event.getPos().getY(), event.getPos().getZ(), 0, 0);
-            dragon.getPhaseManager().setPhase(PhaseList.TAKEOFF);
-            event.getWorld().spawnEntityInWorld(dragon);
+    public static void spawnDragon(BlockEvent.EntityPlaceEvent event) {
+        if (event.getPlacedBlock().getBlock() == Blocks.DRAGON_EGG) {
+            event.getLevel().removeBlock(event.getPos(), false); // false = no flags
+            EnderDragon dragon = EntityType.ENDER_DRAGON.create(event.getEntity().getLevel());
+            dragon.moveTo(event.getPos(), 0, 0);
+            dragon.getPhaseManager().setPhase(EnderDragonPhase.TAKEOFF);
+            event.getLevel().addFreshEntity(dragon);
         }
     }
+
+    // spawn squid
+    // ================================
+
+    // @SubscribeEvent
+    // public static void spawnSquid(BlockEvent.EntityPlaceEvent event) {
+    //     if (event.getPlacedBlock() == Blocks.SPONGE.defaultBlockState()) {
+    //         event.getLevel().removeBlock(event.getPos(), false); // false = no flags
+    //         Squid squid = EntityType.SQUID.create(event.getEntity().level);
+    //         squid.moveTo(event.getPos(), 0, 0);
+    //         event.getLevel().addFreshEntity(squid);
+    //     }
+    // }
+
+    // position offset
+    // ================================
+
+    // @SubscribeEvent
+    // public static void spawnDragon(BlockEvent.EntityPlaceEvent event) {
+    //     if (event.getPlacedBlock() == Blocks.DRAGON_EGG.defaultBlockState()) {
+    //         event.getLevel().removeBlock(event.getPos(), false); // false = no flags
+    //         EnderDragon dragon = EntityType.ENDER_DRAGON.create(event.getEntity().level);
+    //         dragon.moveTo(event.getPos().above(5), 0, 0);
+    //         dragon.getPhaseManager().setPhase(EnderDragonPhase.TAKEOFF);
+    //         event.getLevel().addFreshEntity(dragon);
+    //     }
+    // }
 }
