+++ workshops/minecraft/src/main/java/org/devoxx4kids/forge/mods/SkeletonWar.java
@@ -1,54 +1,51 @@
 package org.devoxx4kids.forge.mods;
 
+import net.minecraft.util.RandomSource;
+import net.minecraft.world.entity.EquipmentSlot;
+import net.minecraft.world.entity.ai.goal.target.NearestAttackableTargetGoal;
+import net.minecraft.world.entity.monster.AbstractSkeleton;
+import net.minecraft.world.item.Item;
+import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.item.Items;
+import net.minecraftforge.event.entity.EntityJoinLevelEvent;
+import net.minecraftforge.eventbus.api.SubscribeEvent;
+import net.minecraftforge.fml.common.Mod;
+
 import java.util.Arrays;
 import java.util.List;
-import java.util.Random;
 
-import net.minecraft.entity.ai.EntityAINearestAttackableTarget;
-import net.minecraft.entity.monster.EntityMob;
-import net.minecraft.entity.monster.EntitySkeleton;
-import net.minecraft.init.Items;
-import net.minecraft.inventory.EntityEquipmentSlot;
-import net.minecraft.item.ItemArmor;
-import net.minecraft.item.ItemStack;
-import net.minecraftforge.event.entity.EntityJoinWorldEvent;
-import net.minecraftforge.fml.common.eventhandler.SubscribeEvent;
-
+@Mod.EventBusSubscriber(modid = MainMod.MODID)
 public class SkeletonWar {
+    static List<Item> helmets = Arrays.asList(Items.LEATHER_HELMET,
+            Items.CHAINMAIL_HELMET, Items.GOLDEN_HELMET, Items.IRON_HELMET,
+            Items.DIAMOND_HELMET, Items.NETHERITE_HELMET, Items.TURTLE_HELMET);
+    static List<Item> chestplates = Arrays.asList(Items.LEATHER_CHESTPLATE,
+            Items.CHAINMAIL_CHESTPLATE, Items.GOLDEN_CHESTPLATE, Items.IRON_CHESTPLATE,
+            Items.DIAMOND_CHESTPLATE, Items.NETHERITE_CHESTPLATE);
+    static List<Item> leggings = Arrays.asList(Items.LEATHER_LEGGINGS,
+            Items.CHAINMAIL_LEGGINGS, Items.GOLDEN_LEGGINGS, Items.IRON_LEGGINGS,
+            Items.DIAMOND_LEGGINGS, Items.NETHERITE_LEGGINGS);
+    static List<Item> boots = Arrays.asList(Items.LEATHER_BOOTS,
+            Items.CHAINMAIL_BOOTS, Items.GOLDEN_BOOTS, Items.IRON_BOOTS,
+            Items.DIAMOND_BOOTS, Items.NETHERITE_BOOTS);
 
     @SubscribeEvent
-    public void makeWarNotPeace(EntityJoinWorldEvent event) {
-        if (!(event.getEntity() instanceof EntitySkeleton)) {
+    public static void makeWarNotPeace(EntityJoinLevelEvent event) {
+        if (!(event.getEntity() instanceof AbstractSkeleton skeleton)) {
             return;
         }
-        EntitySkeleton skeleton = (EntitySkeleton) event.getEntity();
 
-        List<ItemArmor> helmets = Arrays.asList(Items.LEATHER_HELMET,
-                Items.CHAINMAIL_HELMET, Items.GOLDEN_HELMET, Items.IRON_HELMET,
-                Items.DIAMOND_HELMET);
-        List<ItemArmor> chestplates = Arrays.asList(Items.LEATHER_CHESTPLATE,
-                Items.CHAINMAIL_CHESTPLATE, Items.GOLDEN_CHESTPLATE,
-                Items.IRON_CHESTPLATE, Items.DIAMOND_CHESTPLATE);
-        List<ItemArmor> leggings = Arrays.asList(Items.LEATHER_LEGGINGS,
-                Items.CHAINMAIL_LEGGINGS, Items.GOLDEN_LEGGINGS,
-                Items.IRON_LEGGINGS, Items.DIAMOND_LEGGINGS);
-        List<ItemArmor> boots = Arrays.asList(Items.LEATHER_BOOTS,
-                Items.CHAINMAIL_BOOTS, Items.GOLDEN_BOOTS, Items.IRON_BOOTS,
-                Items.DIAMOND_BOOTS);
+        skeleton.targetSelector.addGoal(3, new NearestAttackableTargetGoal<>(skeleton, AbstractSkeleton.class, true));
 
-        skeleton.targetTasks.addTask(3, new EntityAINearestAttackableTarget(
-                skeleton, EntitySkeleton.class, true));
-        skeleton.canPickUpLoot();
-
-        Random random = new Random();
-        skeleton.setItemStackToSlot(EntityEquipmentSlot.HEAD, new ItemStack(
-                helmets.get(random.nextInt(5))));
-        skeleton.setItemStackToSlot(EntityEquipmentSlot.CHEST, new ItemStack(
-                chestplates.get(random.nextInt(5))));
-        skeleton.setItemStackToSlot(EntityEquipmentSlot.LEGS, new ItemStack(
-                leggings.get(random.nextInt(5))));
-        skeleton.setItemStackToSlot(EntityEquipmentSlot.FEET,
-                new ItemStack(boots.get(random.nextInt(5))));
+        RandomSource random = skeleton.level.random;
+        skeleton.setItemSlot(EquipmentSlot.HEAD, new ItemStack(helmets.get(random.nextInt(6))));
+        skeleton.setItemSlot(EquipmentSlot.CHEST, new ItemStack(chestplates.get(random.nextInt(5))));
+        skeleton.setItemSlot(EquipmentSlot.LEGS, new ItemStack(leggings.get(random.nextInt(5))));
+        skeleton.setItemSlot(EquipmentSlot.FEET, new ItemStack(boots.get(random.nextInt(5))));
     }
 
+    // variations can include:
+    // - changing the items in the lists (e.g. adding pumpkins to helmets list)
+    // - changing target mob
+    // - changing priority of goal (player is priority 2, iron golems and turtles are priority 3)
 }
