+++ workshops/minecraft/src/main/resources/pack.mcmeta
@@ -1,0 +1,8 @@
+{
+    "pack": {
+        "description": "examplemod resources",
+        "pack_format": 9,
+        "forge:resource_pack_format": 9,
+        "forge:data_pack_format": 10
+    }
+}
